requirements.txt
# Dependencies for Ship Detection Processing
numpy>=1.20.0
scipy>=1.7.0
matplotlib>=3.4.0
sarpy>=1.4.0
psutil>=5.9.0
dask>=2023.3.0
cupy-cuda11x>=12.0.0  # Optional: For GPU acceleration, use appropriate version for your CUDA
PyWavelets>=1.4.1     # For wavelet transforms
opencv-python>=4.7.0  # For additional image processing
tqdm>=4.65.0          # For progress bars
